% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_call_wrappers.R
\name{build_clique_db_db}
\alias{build_clique_db_db}
\title{Clique DB function wrapper}
\usage{
build_clique_db_db(ppi_name, db_folder, db_name, con)
}
\arguments{
\item{ppi_name}{Name of the PPI network in the database}

\item{db_folder}{A directory where the database will be stored. 
Please not that tilde expension does not work.}

\item{db_name}{The name of the database. File suffix ".sqlite" will be appended}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Clique DB function wrapper
}
\details{
Builds a clique db for all maximal cliques in the PPI network and stores the location in the database,
wrapper for \code{\link[MODifieR]{build_clique_db}}
}
