% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_call_wrappers.R
\name{clique_sum_db}
\alias{clique_sum_db}
\title{Clique SuM function wrapper}
\usage{
clique_sum_db(
  input_name,
  ppi_name,
  n_iterations = 1e+05,
  to_db = TRUE,
  folder = NULL,
  clique_significance = 0.01,
  min_clique_size = 2,
  multiple_cores = T,
  n_cores = 4,
  module_name,
  con
)
}
\arguments{
\item{input_name}{Name of the input object in the database}

\item{ppi_name}{Name of the PPI network in the database}

\item{n_iterations}{Number of iterations to be performed for the permutation based p-value}

\item{to_db}{Save result in database, default is TRUE}

\item{folder}{If not saved in database, folder where to save the module as .rds}

\item{clique_significance}{p-value for cliques to be considered significant}

\item{min_clique_size}{Minimal size for cliques}

\item{multiple_cores}{Parallel process using multiple cores?}

\item{n_cores}{Number of cores to use}

\item{module_name}{Unique name for the module object when stored in database}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Clique SuM function wrapper
}
\details{
Retrieves a MODifieR input object from database by
input_name and runs \code{\link[MODifieR]{clique_sum_permutation}} method.
}
