% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to database}
\usage{
connect_to_db(db)
}
\arguments{
\item{db}{Location of the database}
}
\value{
Returns a database connection object to be used in all database related functions.
}
\description{
Connect to database
}
\details{
Connects to an SQLite database. If the database does not exist, an empty database will be created.
}
