% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_call_wrappers.R
\name{correlation_clique_db}
\alias{correlation_clique_db}
\title{correlation_clique function wrapper}
\usage{
correlation_clique_db(input_name, ppi_name, frequency_cutoff = 0.5,
  to_db = TRUE, folder = NULL, fraction_of_interactions = 0.4,
  iteration = 50, clique_significance = 0.01, deg_cutoff = 0.05,
  multiple_cores = FALSE, n_cores = 3, module_name, con)
}
\arguments{
\item{input_name}{Name of the input object in the database}

\item{ppi_name}{Name of the PPI network in the database}

\item{frequency_cutoff}{Fraction of the number of times a gene should be present in it
iterations. Default is 0.5, meaning 50 procent of all iterations}

\item{to_db}{Save result in database, default is TRUE}

\item{folder}{If not saved in database, folder where to save the module as .rds}

\item{fraction_of_interactions}{Fraction of interactions from the original network that will be used
in each iteration}

\item{iteration}{Number of iterations to be performed}

\item{clique_significance}{Cutoff for Fisher exact test for cliques}

\item{deg_cutoff}{p-value cutoff for differentialy expressed genes}

\item{multiple_cores}{parallelize iterations using number of cores on system -1?}

\item{n_cores}{Number of cores to use}

\item{module_name}{Unique name for the module object when stored in database}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
correlation_clique function wrapper
}
\details{
Retrieves a MODifieR input object from database by
input_name and runs \code{\link[MODifieR]{correlation_clique}} method.
}
