% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_objects.R
\name{delete_input_object}
\alias{delete_input_object}
\title{Delete an input object from the database}
\usage{
delete_input_object(input_name, delete_module_objects = FALSE, con)
}
\arguments{
\item{input_name}{Name of the input object in the database}

\item{delete_module_objects}{Also delete all module objects inferred with this input object? D
efault is FALSE}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Delete an input object from the database
}
