% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_objects.R
\name{delete_module_object}
\alias{delete_module_object}
\title{Delete a module object from database by module_name}
\usage{
delete_module_object(module_name, con)
}
\arguments{
\item{module_name}{Unique name for the module object when stored in database}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Delete a module object from database by module_name
}
