% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_call_wrappers.R
\name{diamond_db}
\alias{diamond_db}
\title{diamond function wrapper}
\usage{
diamond_db(input_name, ppi_name, deg_cutoff = 0.05, to_db = TRUE,
  folder = NULL, n_output_genes = 200, seed_weight = 10,
  include_seed = F, module_name, con)
}
\arguments{
\item{input_name}{Name of the input object in the database}

\item{ppi_name}{Name of the PPI network in the database}

\item{deg_cutoff}{p-value cutoff for differentialy expressed genes}

\item{to_db}{Save result in database, default is TRUE}

\item{folder}{If not saved in database, folder where to save the module as .rds}

\item{n_output_genes}{maximum number of genes to be included in the final module}

\item{seed_weight}{Numeric additional parameter to assign weight for the seed genes}

\item{include_seed}{Logical TRUE/FALSE for inclusion of seed genes in the output module}

\item{module_name}{Unique name for the module object when stored in database}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
diamond function wrapper
}
\details{
Retrieves a MODifieR input object from database by
input_name and runs \code{\link[MODifieR]{diamond}} method.
}
