% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_call_wrappers.R
\name{diffcoex_db}
\alias{diffcoex_db}
\title{DiffCoEx function wrapper}
\usage{
diffcoex_db(input_name, beta = NULL, cor_method = "spearman",
  to_db = TRUE, folder = NULL, cluster_method = "average",
  cuttree_method = "hybrid", cut_height = 0.996, deepSplit = 0,
  pamRespectsDendro = FALSE, minClusterSize = 20, cutHeight = 0.2,
  pval_cutoff = 0.05, module_name, con)
}
\arguments{
\item{input_name}{Name of the input object in the database}

\item{beta}{User-defined soft thresholding power
For method=="tree" it defaults to 0.99. For method=="hybrid" it
defaults to 99% of the range between the 5th percentile and the
 maximum of the joining heights on the dendrogram.}

\item{cor_method}{a character string indicating which correlation coefficient (or covariance)
is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{to_db}{Save result in database, default is TRUE}

\item{folder}{If not saved in database, folder where to save the module as .rds}

\item{cluster_method}{the agglomeration method to be used.
This should be (an unambiguous abbreviation of) one of "ward",
"single", "complete", "average", "mcquitty", "median" or "centroid".
This applies to hierachical clustering.}

\item{cuttree_method}{Chooses the method to use. Recognized values are "hybrid" and "tree".}

\item{cut_height}{Maximum joining heights that will be considered.}

\item{deepSplit}{For method "hybrid", can be either logical or integer in the range 0 to 4. For method
"tree", must be logical. In both cases, provides a rough control over sensitivity to cluster splitting.
The higher the value (or if \code{TRUE}), the more and smaller clusters will be produced. For the
"hybrid" method, a finer control can be achieved via \code{maxCoreScatter} and \code{minGap} below.}

\item{pamRespectsDendro}{Logical, only used for method "hybrid". If \code{TRUE}, the PAM stage will
respect the dendrogram in the sense that objects and small clusters will only be assigned to clusters
that belong to the same branch that the objects or small clusters being assigned belong to.
}

\item{minClusterSize}{Minimum cluster size.  }

\item{cutHeight}{Maximum joining heights that will be considered. For \code{method=="tree"} it
defaults to 0.99. For \code{method=="hybrid"} it defaults to 99\% of the range between the 5th
percentile and the maximum of the joining heights on the dendrogram.}

\item{pval_cutoff}{The p-value cutoff to be used for significant co-expression modules (colors)}

\item{module_name}{Unique name for the module object when stored in database}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
DiffCoEx function wrapper
}
\details{
Retrieves a MODifieR input object from database by
input_name and runs \code{\link[MODifieR]{diffcoex}} method.
}
