% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{disconnect_db}
\alias{disconnect_db}
\title{Disconnect from database}
\usage{
disconnect_db(con)
}
\arguments{
\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Disconnect from database
}
