% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_networks.R
\name{get_available_db_networks}
\alias{get_available_db_networks}
\title{Get all names from PPI networks present in the database}
\usage{
get_available_db_networks(con)
}
\arguments{
\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Get all names from PPI networks present in the database
}
