% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_objects.R
\name{get_available_input_objects}
\alias{get_available_input_objects}
\title{Get all names from input objects present in the database}
\usage{
get_available_input_objects(con)
}
\arguments{
\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Get all names from input objects present in the database
}
