% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_objects.R
\name{get_available_module_objects}
\alias{get_available_module_objects}
\title{Get the names of all module objects in the database}
\usage{
get_available_module_objects(con)
}
\arguments{
\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Get the names of all module objects in the database
}
