% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_networks.R
\name{get_available_networks}
\alias{get_available_networks}
\title{Get the names of all PPI networks available in database}
\usage{
get_available_networks(con)
}
\arguments{
\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Get the names of all PPI networks available in database
}
