% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_genes}
\alias{get_unique_genes}
\title{Get all unique genes present in the database}
\usage{
get_unique_genes(con)
}
\arguments{
\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\value{
A vector containing all unique genes present in the database, both in input and module objects
}
\description{
Get all unique genes present in the database
}
