% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_call_wrappers.R
\name{mcode_db}
\alias{mcode_db}
\title{MCODE function wrapper}
\usage{
mcode_db(input_name, ppi_name, to_db = TRUE, folder = NULL,
  hierarchy = 1, vwp = 0.5, haircut = F, fluff = F, fdt = 0.8,
  loops = T, deg_cutoff = 0.05, module_cutoff = 3.5, module_name,
  con)
}
\arguments{
\item{input_name}{Name of the input object in the database}

\item{ppi_name}{Name of the PPI network in the database}

\item{to_db}{Save result in database, default is TRUE}

\item{folder}{If not saved in database, folder where to save the module as .rds}

\item{hierarchy}{This parameter indicates how many hierarchy are included in the network, currently it can be \code{0}, \code{1} or \code{2}. Default value is \code{1}.}

\item{vwp}{Vertex weight percentage. Default value is 0.5.}

\item{haircut}{Boolean value, whether to remove singly-connected nodes from clusters (TRUE) or not (FALSE).}

\item{fluff}{Boolean value, whether to spand cluster cores by one neighbour shell outwards (TRUE) or not (FALSE).}

\item{fdt}{Cluster density cutoff. Default value is 0.8.}

\item{loops}{Boolean value, whether to include self-loops (TRUE) or not (FALSE).}

\item{deg_cutoff}{p-value cutoff for differentialy expressed genes}

\item{module_cutoff}{Minimal score for a module to be returned}

\item{module_name}{Unique name for the module object when stored in database}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
MCODE function wrapper
}
\details{
Retrieves a MODifieR input object from database by input_name and runs \code{\link[MODifieR]{mcode}} method.
}
