% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_call_wrappers.R
\name{moda_db}
\alias{moda_db}
\title{moda function wrapper}
\usage{
moda_db(
  input_name,
  cutmethod = "Density",
  to_db = TRUE,
  folder = NULL,
  group_of_interest,
  specificTheta = 0.1,
  conservedTheta = 0.1,
  module_name,
  con
)
}
\arguments{
\item{input_name}{Name of the input object in the database}

\item{cutmethod}{cutting the dendrogram based on maximal average Density 
or Modularity}

\item{to_db}{Save result in database, default is TRUE}

\item{folder}{If not saved in database, folder where to save the module as .rds}

\item{group_of_interest}{Numerical value denoting which group contains the condition of interest (1 or 2)}

\item{specificTheta}{the threshold to define min(s)+specificTheta, 
less than which is considered as condition specific module. 
s is the sums of rows in Jaccard index matrix. See supplementary file.}

\item{conservedTheta}{The threshold to define max(s)-conservedTheta, 
greater than which is considered as condition conserved module. 
s is the sums of rows in Jaccard index matrix. See supplementary file.}

\item{module_name}{Unique name for the module object when stored in database}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
moda function wrapper
}
\details{
Retrieves a MODifieR input object from database by
input_name and runs \code{\link[MODifieR]{moda}} method.
}
