% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_objects.R
\name{MODifieR_input_from_db}
\alias{MODifieR_input_from_db}
\title{Retrieve an input object from the database by input_name}
\usage{
MODifieR_input_from_db(input_name, con)
}
\arguments{
\item{input_name}{Name of the input object in the database}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Retrieve an input object from the database by input_name
}
