% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_objects.R
\name{MODifieR_module_from_db}
\alias{MODifieR_module_from_db}
\title{Retrieve an module object from the database by module_name}
\usage{
MODifieR_module_from_db(module_name, con)
}
\arguments{
\item{module_name}{Unique name for the module object when stored in database}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Retrieve an module object from the database by module_name
}
