% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MODifieR_object_to_db}
\alias{MODifieR_object_to_db}
\title{Stores a MODifieR object in the database}
\usage{
MODifieR_object_to_db(MODifieR_object, object_name, con)
}
\arguments{
\item{MODifieR_object}{Any MODifieR input or module object}

\item{object_name}{Unique name for the object in the database}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Stores a MODifieR object in the database
}
