% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_call_wrappers.R
\name{modulediscoverer_db}
\alias{modulediscoverer_db}
\title{Module Discoverer function wrapper}
\usage{
modulediscoverer_db(
  input_name,
  ppi_name,
  permutations = 10000,
  to_db = TRUE,
  folder = NULL,
  deg_cutoff = 0.05,
  repeats = 15,
  clique_cutoff = 0.01,
  n_cores,
  module_name,
  con
)
}
\arguments{
\item{input_name}{Name of the input object in the database}

\item{ppi_name}{Name of the PPI network in the database}

\item{permutations}{Number of permutations to perform to identify the community structure}

\item{to_db}{Save result in database, default is TRUE}

\item{folder}{If not saved in database, folder where to save the module as .rds}

\item{deg_cutoff}{p-value cutoff for differentialy expressed genes}

\item{repeats}{Number of times the algorithm is repeated}

\item{clique_cutoff}{cutoff pvalue for significant cliques}

\item{module_name}{Unique name for the module object when stored in database}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Module Discoverer function wrapper
}
\details{
Retrieves a MODifieR input object from database by
input_name and runs \code{\link[MODifieR]{modulediscoverer}} method.
}
