% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_networks.R
\name{ppi_network_from_db}
\alias{ppi_network_from_db}
\title{Retrieve a PPI network from database by ppi_name}
\usage{
ppi_network_from_db(ppi_name, con)
}
\arguments{
\item{ppi_name}{Name of the PPI network in the database}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Retrieve a PPI network from database by ppi_name
}
