% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_networks.R
\name{ppi_network_to_db}
\alias{ppi_network_to_db}
\title{Store a PPI network in the database}
\usage{
ppi_network_to_db(ppi_network, ppi_name, con)
}
\arguments{
\item{ppi_network}{A network as a dataframe where the first 2 columns are the interactions}

\item{ppi_name}{Name of the PPI network in the database}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
Store a PPI network in the database
}
