% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_call_wrappers.R
\name{wgcna_db}
\alias{wgcna_db}
\title{wgcna function wrapper}
\usage{
wgcna_db(input_name, group_of_interest, minModuleSize = 30,
  deepSplit = 2, to_db = TRUE, folder = NULL,
  pamRespectsDendro = FALSE, mergeCutHeight = 0.1,
  numericLabels = TRUE, pval_cutoff = 0.05, corType = "bicor",
  maxBlockSize = 10000, TOMType = "signed", saveTOMs = TRUE,
  maxPOutliers = 0.1, module_name, con)
}
\arguments{
\item{input_name}{Name of the input object in the database}

\item{group_of_interest}{Numerical value denoting which group contains the condition of interest (1 or 2)}

\item{minModuleSize}{ minimum module size for module detection. See
\code{\link[dynamicTreeCut]{cutreeDynamic}} for more details.  }

\item{deepSplit}{ integer value between 0 and 4. Provides a simplified control over how sensitive
module detection should be to module splitting, with 0 least and 4 most sensitive. See
\code{\link[dynamicTreeCut]{cutreeDynamic}} for more details. }

\item{to_db}{Save result in database, default is TRUE}

\item{folder}{If not saved in database, folder where to save the module as .rds}

\item{pamRespectsDendro}{Logical, only used when \code{pamStage} is \code{TRUE}. 
If \code{TRUE}, the PAM stage will
respect the dendrogram in the sense an object can be PAM-assigned only to clusters that lie below it on
the branch that the object is merged into. 
     See \code{\link[dynamicTreeCut]{cutreeDynamic}} for more details. }

\item{mergeCutHeight}{ dendrogram cut height for module merging. }

\item{numericLabels}{ logical: should the returned modules be labeled by colors (\code{FALSE}), or by
numbers (\code{TRUE})? }

\item{pval_cutoff}{The p-value cutoff to be used for significant co-expression modules (colors)}

\item{corType}{character string specifying the correlation to be used. Allowed values are (unique
abbreviations of) \code{"pearson"} and \code{"bicor"}, corresponding to Pearson and bidweight
midcorrelation, respectively. Missing values are handled using the \code{pairwise.complete.obs} option. }

\item{maxBlockSize}{integer giving maximum block size for module detection. Ignored if \code{blocks}
above is non-NULL. Otherwise, if the number of genes in \code{datExpr} exceeds \code{maxBlockSize}, genes
will be pre-clustered into blocks whose size should not exceed \code{maxBlockSize}. }

\item{TOMType}{ one of \code{"none"}, \code{"unsigned"}, \code{"signed"}, \code{"signed Nowick"},
\code{"unsigned 2"}, \code{"signed 2"} and \code{"signed Nowick 2"}. If \code{"none"}, adjacency
will be used for clustering. See \code{\link{TOMsimilarityFromExpr}} for details.}

\item{saveTOMs}{ logical: should the consensus topological overlap matrices for each block be saved
and returned?  }

\item{maxPOutliers}{ only used for \code{corType=="bicor"}. Specifies the maximum percentile of data 
that can be considered outliers on either 
side of the median separately. For each side of the median, if
higher percentile than \code{maxPOutliers} is considered an outlier by the weight function based on
\code{9*mad(x)}, the width of the weight function is increased such that the percentile of outliers on
that side of the median equals \code{maxPOutliers}. Using \code{maxPOutliers=1} will effectively disable
all weight function broadening; using \code{maxPOutliers=0} will give results that are quite similar (but
not equal to) Pearson correlation. }

\item{module_name}{Unique name for the module object when stored in database}

\item{con}{Database connection obtained with \code{\link{connect_to_db}} function}
}
\description{
wgcna function wrapper
}
\details{
Retrieves a MODifieR input object from database by
input_name and runs \code{\link[MODifieR]{wgcna}} method.
}
